refactor: change models file names
