+++ src/album/album.controller.ts
@@ -14,7 +14,7 @@ import {
 import { AlbumService } from './album.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
-import { Album } from '../models/album.entity';
+import { Album } from '../models/album';
 import { ParseUUIDPipe } from '@nestjs/common';
 import { ApiBody, ApiOperation, ApiParam } from '@nestjs/swagger';
