+++ src/album/album.service.ts
@@ -2,7 +2,7 @@
 
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { AlbumRepository } from './album.repository';
-import { Album } from '../models/album.entity';
+import { Album } from '../models/album';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
 import { TrackService } from 'src/track/track.service';
