+++ src/artist/artist.controller.ts
@@ -12,7 +12,7 @@ import {
   ValidationPipe,
 } from '@nestjs/common';
 import { ArtistService } from './artist.service';
-import { Artist } from '../models/artist.entity';
+import { Artist } from '../models/artist';
 import { ParseUUIDPipe } from '@nestjs/common';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
