+++ src/artist/artist.repository.ts
@@ -1,4 +1,4 @@
 import { BaseRepository } from '../repositories/in-memory-db/base.repository';
-import { Artist } from '../models/artist.entity';
+import { Artist } from '../models/artist';
 
 export class ArtistRepository extends BaseRepository<Artist> {}
