+++ src/models/artist.ts
